Placing heartbeatManager in separate file
